requirements.txt
pycryptodome
websocket
websocket-client
